+++ Backend-farmlogitech/Monitoring/Domain/Model/Commands/Messages/CreateMessageCommand.cs
@@ -1,3 +1,3 @@
 namespace Backend_farmlogitech.Monitoring.Domain.Model.Commands.Messages;
 
-public record CreateMessageCommand(string description, long collaboratorId, long farmerId, long transmitterId);
+public record CreateMessageCommand(string description, int collaboratorId, int farmerId, int transmitterId);
