+++ Backend-farmlogitech/Monitoring/Domain/Model/Queries/Message/GetAllMessagesByCollaboratorId.cs
@@ -1,3 +1,3 @@
 namespace Backend_farmlogitech.Monitoring.Domain.Model.Queries.Message;
 
-public record GetAllMessagesByCollaboratorId(long collaboratorId);
+public record GetAllMessagesByCollaboratorId(int collaboratorId);
