+++ Backend-farmlogitech/Monitoring/Domain/Model/Queries/Message/GetAllMessagesByCollaboratorIdAndFarmerId.cs
@@ -1,3 +1,3 @@
 ﻿namespace Backend_farmlogitech.Monitoring.Domain.Model.Queries.Message;
 
-public record GetAllMessagesByCollaboratorIdAndFarmerId(long userId, long transmitterId);
+public record GetAllMessagesByCollaboratorIdAndFarmerId(int userId, int transmitterId);
