+++ Backend-farmlogitech/Monitoring/Interfaces/REST/Resources/Messages/CreateMessageResource.cs
@@ -1,3 +1,3 @@
 namespace Backend_farmlogitech.Monitoring.Interfaces.REST.Resources.Messages;
 
-public record CreateMessageResource(string description, long collaboratorId, long farmerId, long transmitterId);
+public record CreateMessageResource(string description, int collaboratorId, int farmerId, int transmitterId);
