+++ Backend-farmlogitech/Monitoring/Interfaces/REST/Resources/Messages/MessageResource.cs
@@ -1,3 +1,3 @@
 namespace Backend_farmlogitech.Monitoring.Interfaces.REST.Resources.Messages;
 
-public record MessageResource(long id, string description, long collaboratorId, long farmerId, long transmitterId);
+public record MessageResource(int id, string description, int collaboratorId, int farmerId, int transmitterId);
